+++ HISTORY.md
@@ -1,6 +1,8 @@
 V 1.0.2 March 17, 2016
 ============================
 * Postmedia.Files.ClassFilename Sniff added, checks for ClassFilenameNotCamelCaps
+* Postmedia.Classes.ValidClassName Sniff added, enforces CamelCaps for the class name to match the filename
+* Added method to Postmedia.Files.ClassFilename Sniff that checks for two or more classes present.
 
 V 1.0.1 March 11, 2016
 ============================